\name{chamois}
\docType{data}
\alias{chamois}
\title{Location of Chamois Groups in the Chartreuse Mountains}
\description{
  This data set describes the habitat use and availability by the
  chamois of the Chartreuse mountains (Isère, France), in 1992 and 1997.
  These data have been gathered during the hunting season (Fall).
}
\usage{
data(chamois)
}
\details{
  The object \code{chamois} is a list containing the following
  components:
  \describe{
    \item{\code{locs}}{A data frame containing the x and y coordinates
      of 198 chamois groups.}
    \item{\code{map}}{a map of class \code{kasc} describing the
      vegetation (Forest or Open areas), the
      distance from the ecotone Open/Forest and the slopes on the
      area.}
    }
}
\references{
  Fédération Départementale des Chasseurs de l'Isère,
  65 av Jean Jaurès, 38320 Eybens. France.
}
\keyword{datasets}
