\name{puechdesIII}
\alias{puechdesIII}
\docType{data}
\title{Habitat Selection by the Wild Boar at Puechabon}
\description{
This data set contains two data frames describing the use and the
availability of 3 elevation classes for 6 wild boars (\emph{Sus scrofa}
L.) monitored using radio-tracking at Puechabon (South of France).
These data have been collected by Daniel Maillard (Office national de la
chasse et de la faune sauvage).
}
\usage{data(puechdesIII)}
\details{
  The list \code{puechdesIII} has two components: \cr
  The data frame \code{used} describes the number of telemetry
  relocations  for each of the 6 animals in each of the three elevation
  classes. \cr
  The data frame \code{available} describes a sample 
  of random points placed in the areas available to these wild boars 
  (a buffer area of 200 m around the relocations).
}
\source{
Maillard, D. (1996) \emph{Occupation et utilisation de la garrigue et du
vignoble mediterraneens par le Sanglier}.
Universite d'Aix-Marseille III:  PhD thesis.
}
\keyword{datasets}
