\name{rotxy}
\alias{rotxy}
\alias{.buffer.point.unic}
\alias{profilehab}
\alias{[.kasc}
\alias{[[.kasc}
\title{ Internal Functions for Package 'adehabitat' }
\description{
Internal functions for the package \code{adehabitat}.
}
\usage{
}

\details{
These are not to be called by the user.
}
\keyword{internal}
