\name{chamois}
\docType{data}
\alias{chamois}
\title{Location of Chamois Groups in the Chartreuse Mountains}
\description{
  This data set describes the habitat use and availability by the
  chamois of the Chartreuse mountains (Isere, France), in 1992 and 1997.
  These data have been gathered during the hunting season (Fall).
}
\usage{
data(chamois)
}
\details{
  The object \code{chamois} is a list containing the following
  components:\cr
  \code{locs} is a data frame containing the x and y coordinates
  of 198 chamois groups.\cr
  \code{map} is a map of class \code{kasc} describing the
  vegetation (Forest or Open areas), the distance from the ecotone
  Open/Forest and the slopes on the area.
}
\references{
  Federation Departementale des Chasseurs de l'Isere,
  65 av Jean Jaures, 38320 Eybens. France.
}
\keyword{datasets}
