\name{rotxy}
\alias{rotxy}
\alias{.buffer.point.unic}
\alias{profilehab}
\alias{[.kasc}
\alias{[[.kasc}
\alias{NNCH.iso.index}
\alias{vectorize.parameter}
\alias{noholes.poly}
\alias{autourna}
\alias{.corrXY}
\alias{.convtime}
\title{Internal Functions for Package 'adehabitat'}
\description{
Internal functions for the package \code{adehabitat}.
}

\details{
These are not to be called by the user.
}
\keyword{internal}
